Block versionin - uint32 in schema - type changed to long; using global version for all blocks now. + folderId bugfix

+++ PhraserManager/src/main/java/com/phraser/db/Block.java
@@ -79,7 +79,7 @@ public interface Block {
     return blockType();
   }
 
-  default int getVersion() {
+  default long getVersion() {
     return storeBlock().version();
   }
